+++ lib/src/version_helper.dart
@@ -2,4 +2,5 @@
 
 // The version is now hardcoded here.
 // Remember to change this manually when you release a new version.
-const String kDigCliVersion = '1.3.0';
+const String kDigCliVersion = '1.3.1';
+